added password gen pref ; no finish() in onPause()
No reason to kill the activity as long as clearSecrets() is always run
in a disciplined way. So now we can onPause() and then onResume() and
the login screen will show again. Timer is also stopped and re-started
as appropriate.

Preferences screen that lets us set the generated password length.
Perhaps we'll add one for allowed special characters as well? Hmm.